+++ CHANGELOG.md
@@ -15,9 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added @nylira/vue-page-header as `tmPageHeader`
 - Added @nylira/vue-notifications as `tmNotifications`
 - Added @nylira/vue-countdown as `tmCountdown`
-
-### Changes
-
+- Added @nylira/vue-key-value(s) as `tmKeyValue(s)`
 - Added @nylira/vue-field as `tmField`
 
 ### Changes